+++ MonoGame.Framework/Graphics/PackedVector/Rgba1010102.cs
@@ -60,10 +60,10 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
         public Vector4 ToVector4()
         {
             return new Vector4(
-                (float) (((packedValue >> 0) & 0x03FF) / 1023.0f),
-                (float) (((packedValue >> 10) & 0x03FF) / 1023.0f),
-                (float) (((packedValue >> 20) & 0x03FF) / 1023.0f),
-                (float) (((packedValue >> 30) & 0x03) / 3.0f)
+(((packedValue >> 0) & 0x03FF) / 1023.0f),
+(((packedValue >> 10) & 0x03FF) / 1023.0f),
+(((packedValue >> 20) & 0x03FF) / 1023.0f),
+(((packedValue >> 30) & 0x03) / 3.0f)
             );
         }
 
@@ -83,7 +83,7 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
         /// <returns>True if the object is equal to the packed vector.</returns>
         public override bool Equals(object obj)
         {
-            return (obj is Rgba1010102) && Equals((Rgba1010102) obj);
+            return (obj is Rgba1010102) && Equals((Rgba1010102)obj);
         }
 
         /// <summary>
@@ -126,11 +126,11 @@ namespace Microsoft.Xna.Framework.Graphics.PackedVector
 
         private static uint Pack(float x, float y, float z, float w)
         {
-            return (uint) (
-                (((int) Math.Round(MathHelper.Clamp(x, 0, 1) * 1023.0f) & 0x03FF) << 0) |
-                (((int) Math.Round(MathHelper.Clamp(y, 0, 1) * 1023.0f) & 0x03FF) << 10) |
-                (((int) Math.Round(MathHelper.Clamp(z, 0, 1) * 1023.0f) & 0x03FF) << 20) |
-                (((int) Math.Round(MathHelper.Clamp(w, 0, 1) * 3.0f) & 0x03) << 30)
+            return (uint)(
+                (((int)Math.Round(MathHelper.Clamp(x, 0, 1) * 1023.0f) & 0x03FF) << 0) |
+                (((int)Math.Round(MathHelper.Clamp(y, 0, 1) * 1023.0f) & 0x03FF) << 10) |
+                (((int)Math.Round(MathHelper.Clamp(z, 0, 1) * 1023.0f) & 0x03FF) << 20) |
+                (((int)Math.Round(MathHelper.Clamp(w, 0, 1) * 3.0f) & 0x03) << 30)
             );
         }
     }
